[#14] fix: selector 컴포넌트 section 안으로 이동
